+++ apps/lottery/views.py
@@ -5,67 +5,72 @@ from random import SystemRandom
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required, permission_required
 from django.core.urlresolvers import reverse
-from django.http import Http404
 from django.shortcuts import get_object_or_404, redirect, render
 from django.utils.translation import ugettext as _
+from django.views.decorators.http import require_POST, require_safe
 
 from apps.lottery.models import Lottery, LotteryParticipant, LotteryWinner
+from apps.userprofile.models import User
 
 
+@require_safe
 def details(request, lottery_id):
     lottery = get_object_or_404(Lottery, pk=lottery_id)
     lan = lottery.lan
-    participants = [participant.user for participant in lottery.lotteryparticipant_set.all()]
-    winners_qs = LotteryWinner.objects.filter(lottery=lottery).order_by('-id')
-    winner = None
-    if winners_qs.exists():
-        winner = winners_qs[0]
+    participants = User.objects.filter(lotteryparticipant__lottery=lottery).order_by('username')
+    winners = [w.user for w in LotteryWinner.objects.filter(lottery=lottery).order_by('id').select_related('user')]
+    if winners:
+        last_winner = winners[-1]
+    else:
+        last_winner = None
 
     breadcrumbs = (
         (lan, reverse('lan_details', kwargs={'lan_id': lan.id})),
-        (_(u'Lottery'), ''),
+        (_(u'Competitions'), reverse('competitions_lan_compos', kwargs={'lan_id': lan.id})),
+        (unicode(lottery), ''),
     )
 
-    return render(request, 'lottery/details.html', {'lottery': lottery,
-                  'participants': participants, 'breadcrumbs': breadcrumbs, 'winner': winner})
+    return render(request, 'lottery/details.html', {'breadcrumbs': breadcrumbs, 'lottery': lottery,
+                  'participants': participants, 'winners': winners, 'last_winner': last_winner})
 
 
+@require_POST
 @login_required
 def sign_up(request, lottery_id):
     lottery = get_object_or_404(Lottery, pk=lottery_id)
     if lottery.registration_open and not lottery.is_participating(request.user):
         LotteryParticipant.objects.create(user=request.user, lottery=lottery)
-
     return redirect(lottery)
 
 
+@require_POST
 @login_required
 def sign_off(request, lottery_id):
     lottery = get_object_or_404(Lottery, pk=lottery_id)
     if lottery.registration_open and lottery.is_participating(request.user):
-        participant = LotteryParticipant.objects.get(user=request.user, lottery=lottery)
-        participant.delete()
-
+        LotteryParticipant.objects.filter(user=request.user, lottery=lottery).delete()
     return redirect(lottery)
 
 
-@permission_required('lottery.draw')
-def drawing(request, lottery_id=False):
-    if lottery_id:
-        lottery = get_object_or_404(Lottery, pk=lottery_id)
-    elif Lottery.objects.count() > 0:
-        lottery = Lottery.objects.latest('pk')
-    else:
-        raise Http404(_(u'Lottery not found.'))
-
-    winners = LotteryWinner.objects.filter(lottery=lottery).order_by('-id').all()
-    winner = None
-    if winners:
-        winner = winners[0]
-
-    return render(request, 'lottery/drawing.html', {'lottery': lottery, 'winners': winners, 'winner': winner})
+@require_POST
+@permission_required('lottery.open_close')
+def open_(request, lottery_id):
+    lottery = get_object_or_404(Lottery, pk=lottery_id)
+    lottery.registration_open = True
+    lottery.save()
+    return redirect(lottery)
 
 
+@require_POST
+@permission_required('lottery.open_close')
+def close(request, lottery_id):
+    lottery = get_object_or_404(Lottery, pk=lottery_id)
+    lottery.registration_open = False
+    lottery.save()
+    return redirect(lottery)
+
+
+@require_POST
 @permission_required('lottery.draw')
 def draw(request, lottery_id):
     lottery = get_object_or_404(Lottery, pk=lottery_id)
@@ -73,7 +78,7 @@ def draw(request, lottery_id):
 
     if len(participants) < 1:
         messages.error(request, _(u'No eligible participants.'))
-        return redirect(drawing, lottery_id)
+        return redirect(lottery)
 
     rand = SystemRandom()
     winner_id = rand.randrange(0, len(participants))
@@ -84,4 +89,4 @@ def draw(request, lottery_id):
 
     LotteryWinner.objects.create(user=winner, lottery=lottery)
 
-    return redirect(drawing, lottery_id)
+    return redirect(lottery)